Move utils to service layer
