+++ app/repository/document.py
@@ -24,7 +24,6 @@ from db_client.models.organisation.organisation import Organisation
 from sqlalchemy import func
 from sqlalchemy.orm import Session
 
-from app.core.util import to_cdn_url
 from app.models.document import (
     CollectionOverviewResponse,
     FamilyAndDocumentsResponse,
@@ -36,6 +35,7 @@ from app.models.document import (
 )
 from app.repository.geography import get_geo_subquery
 from app.repository.lookups import doc_type_from_family_document_metadata
+from app.service.util import to_cdn_url
 
 _LOGGER = logging.getLogger(__file__)
